For #201. Add timezone controller instead of using Survey Controller
