+++ web/views/timezone_view.ex
@@ -1,0 +1,9 @@
+defmodule Ask.TimezoneView do
+  use Ask.Web, :view
+
+  def render("index.json", %{timezones: timezones}) do
+    %{
+      timezones: timezones
+    }
+  end
+end
